requirements.txt
fastapi
uvicorn
celery[redis]
redis
python-docx
PyMuPDF
transformers
sentencepiece
torch
boto3
minio
python-multipart
requests

